Death Message Class
Move the data for the death message over to it's own class file, utilizing the event generated by the plugin instead of a direct call.
