+++ src/main/java/me/truedarklord/deathChests/listeners/Death.java
@@ -2,7 +2,6 @@ package me.truedarklord.deathChests.listeners;
 
 import me.truedarklord.deathChests.DeathChests;
 import me.truedarklord.deathChests.events.DeathChestEvent;
-import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
@@ -14,18 +13,12 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.plugin.Plugin;
 
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class Death implements Listener {
 
-    private final Plugin plugin;
-
     public Death(DeathChests plugin) {
-        this.plugin = plugin;
         plugin.getServer().getPluginManager().registerEvents(this, plugin);
     }
 
@@ -41,30 +34,6 @@ public class Death implements Listener {
         while (!drops.isEmpty()) {
             setInvDrops(findEmptyBlock(loc), drops);
         }
-
-        sendDeathMessage(event.getEntity(), loc);
-
-    }
-
-    private void sendDeathMessage(Player player, Location loc) {
-        String msg = plugin.getConfig().getString("Death_Message", "&2You Died at &6%world% X:%x% Y:%y% Z:%z%&2.");
-        msg = msg.replaceAll("%world%", loc.getWorld().getName())
-              .replaceAll("%x%", String.valueOf(loc.getBlockX()))
-             .replaceAll("%y%", String.valueOf(loc.getBlockY()))
-             .replaceAll("%z%", String.valueOf(loc.getBlockZ()));
-        
-        player.sendMessage(colour(msg));
-    }
-
-    private static String colour(String message) {
-        Matcher matcher = Pattern.compile("#[a-fA-F\\d]{6}").matcher(message);
-        StringBuilder sb = new StringBuilder();
-
-        while (matcher.find()) {
-            matcher.appendReplacement(sb, net.md_5.bungee.api.ChatColor.of(matcher.group()).toString());
-        }
-
-        return ChatColor.translateAlternateColorCodes('&', matcher.appendTail(sb).toString());
     }
 
     private Block findEmptyBlock(Location loc) {
